+++ modules/client/src/test/java/org/apache/ignite/client/fakes/FakeIgnite.java
@@ -29,6 +29,7 @@ import org.apache.ignite.internal.replicator.TablePartitionId;
 import org.apache.ignite.internal.sql.engine.QueryProcessor;
 import org.apache.ignite.internal.tx.InternalTransaction;
 import org.apache.ignite.internal.tx.TxState;
+import org.apache.ignite.internal.tx.impl.IgniteTransactionsImpl;
 import org.apache.ignite.lang.IgniteBiTuple;
 import org.apache.ignite.network.ClusterNode;
 import org.apache.ignite.sql.IgniteSql;
@@ -79,91 +80,7 @@ public class FakeIgnite implements Ignite {
     /** {@inheritDoc} */
     @Override
     public IgniteTransactions transactions() {
-        return new IgniteTransactions() {
-            @Override
-            public Transaction begin(TransactionOptions options) {
-                return beginAsync(options).join();
-            }
-
-            @Override
-            public CompletableFuture<Transaction> beginAsync(TransactionOptions options) {
-                return CompletableFuture.completedFuture(new InternalTransaction() {
-                    private final UUID id = UUID.randomUUID();
-
-                    private final HybridTimestamp timestamp = clock.now();
-
-                    @Override
-                    public @NotNull UUID id() {
-                        return id;
-                    }
-
-                    @Override
-                    public IgniteBiTuple<ClusterNode, Long> enlistedNodeAndTerm(TablePartitionId tablePartitionId) {
-                        return null;
-                    }
-
-                    @Override
-                    public TxState state() {
-                        return null;
-                    }
-
-                    @Override
-                    public boolean assignCommitPartition(TablePartitionId tablePartitionId) {
-                        return false;
-                    }
-
-                    @Override
-                    public TablePartitionId commitPartition() {
-                        return null;
-                    }
-
-                    @Override
-                    public IgniteBiTuple<ClusterNode, Long> enlist(
-                            TablePartitionId tablePartitionId,
-                            IgniteBiTuple<ClusterNode, Long> nodeAndTerm) {
-                        return null;
-                    }
-
-                    @Override
-                    public void enlistResultFuture(CompletableFuture<?> resultFuture) {}
-
-                    @Override
-                    public void commit() throws TransactionException {
-
-                    }
-
-                    @Override
-                    public CompletableFuture<Void> commitAsync() {
-                        return CompletableFuture.completedFuture(null);
-                    }
-
-                    @Override
-                    public void rollback() throws TransactionException {
-
-                    }
-
-                    @Override
-                    public CompletableFuture<Void> rollbackAsync() {
-                        return CompletableFuture.completedFuture(null);
-                    }
-
-                    @Override
-                    public boolean isReadOnly() {
-                        return false;
-                    }
-
-                    @Override
-                    public HybridTimestamp readTimestamp() {
-                        return null;
-                    }
-
-                    @Override
-                    public HybridTimestamp startTimestamp() {
-                        return timestamp;
-                    }
-                });
-            }
-        };
+        return new IgniteTransactionsImpl(new FakeTxManager(clock));
     }
 
     /** {@inheritDoc} */
